feat: skip ip integrity check for ipv6, add initial tests setup
