+++ package.json
@@ -31,9 +31,12 @@
     "sqlite3": "^5.0.8"
   },
   "devDependencies": {
+    "@vitejs/plugin-react": "^3.0.1",
     "eslint": "^8.17.0",
     "eslint-config-next": "^12.1.6",
     "eslint-config-prettier": "^8.5.0",
-    "prettier": "2.6.2"
+    "jsdom": "^21.1.0",
+    "prettier": "2.6.2",
+    "vitest": "^0.28.3"
   }
 }
