+++ pages/api/credential-stuffing/authenticate.js
@@ -1,11 +1,5 @@
 import { Sequelize } from 'sequelize';
 import {
-  checkConfidenceScore,
-  checkFreshIdentificationRequest,
-  checkIpAddressIntegrity,
-  checkOriginsIntegrity,
-  CheckResult,
-  checkResultType,
   ensurePostRequest,
   ensureValidRequestIdAndVisitorId,
   getForbiddenResponse,
@@ -15,6 +9,13 @@ import {
   reportSuspiciousActivity,
   sequelize,
 } from '../../../server/server';
+import { CheckResult, checkResultType } from '../../../server/checkResult';
+import {
+  checkConfidenceScore,
+  checkFreshIdentificationRequest,
+  checkIpAddressIntegrity,
+  checkOriginsIntegrity,
+} from '../../../server/checks';
 
 // Defines db model for login attempt.
 export const LoginAttempt = sequelize.define('login-attempt', {
