+++ pages/api/loan-risk/request-loan.js
@@ -1,8 +1,9 @@
 import { loanRiskEndpoint } from '../../../server/loan-risk/loan-risk-endpoint';
 import { LoanRequest } from '../../../server/loan-risk/database';
 import { Op } from 'sequelize';
-import { CheckResult, checkResultType, messageSeverity } from '../../../server/server';
+import { messageSeverity } from '../../../server/server';
 import { calculateLoanValues } from '../../../server/loan-risk/calculate-loan-values';
+import { CheckResult, checkResultType } from '../../../server/checkResult';
 
 /**
  * Validates previous loan requests sent by a given user.
