+++ server/const.js
@@ -1,0 +1,4 @@
+export const IPv4_REGEX = /^((25[0-5]|(2[0-4]|1\d|[1-9]|)\d)\.){3}(25[0-5]|(2[0-4]|1\d|[1-9]|)\d)$/;
+export const ALLOWED_REQUEST_TIMESTAMP_DIFF_MS = 3000;
+// Confidence score thresholds might be different for different scenarios
+export const MIN_CONFIDENCE_SCORE = 0.85;
